Delphi Game Launcher
A basic game / exe launcher made with Delphi 10.4

DISCLAIMER:
If you install the game using the "Delphi-Launcher-SETUP.exe" file, you will need to right click on the shortcut on your desktop, go to properties. Under "Shortcut" click Advanced and check 'Run as administrator' in order for the program to be able to read/write to the data files.